Move yamaha_musiccast base entity to separate module
