+++ homeassistant/components/yamaha_musiccast/select.py
@@ -9,8 +9,9 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import DOMAIN, MusicCastCapabilityEntity, MusicCastDataUpdateCoordinator
+from . import DOMAIN, MusicCastDataUpdateCoordinator
 from .const import TRANSLATION_KEY_MAPPING
+from .entity import MusicCastCapabilityEntity
 
 
 async def async_setup_entry(
